+++ CHANGELOG.md
@@ -98,12 +98,14 @@ And (*as always*) many bugfixes and improvements to existing features!
   - use model vae scale-factor for image width/heigt calculations  
 - **Other**
   - **prompt enhance** add `allura-org/Gemma-3-Glitter-4B` model support  
-  - remove **LDSR**  
-  - remove `api-only` cli option  
+  - **prompt enhance** add `Qwen/Qwen3-4B-Instruct-2507` model support  
+  - **schedulers** add **Flash FlowMatch**  
 - **API**
   - add `/sdapi/v1/checkpoint` POST endpoint to simply load a model  
 - **Refactor**
   - new unified pipeline component loader in `pipelines/generic`  
+  - remove **LDSR**  
+  - remove `api-only` cli option  
 - **Fixes**  
   - refactor legacy processing loop  
   - fix settings components mismatch